Create spatialD.py
An ArcPy batch processing script that searches a workspace for vector dataset, checks to see if it has a spatial reference system defined, and if not, then assigns one. Also provides total for the number of feature classes in the workspace, total number of feature classes with a spatial reference system, and list of names of feature classes without a spatial reference system.